XRDDEV-2187 Deleting unused SignerProxyService. Code is duplicated in SignerProxyFacade
