+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/centralserver/restapi/facade/SignerProxyFacadeImpl.java
@@ -36,10 +36,17 @@ import ee.ria.xroad.signer.protocol.dto.TokenInfo;
 import ee.ria.xroad.signer.protocol.dto.TokenInfoAndKeyId;
 import ee.ria.xroad.signer.protocol.message.CertificateRequestFormat;
 
+import akka.actor.ActorSystem;
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
 import lombok.extern.slf4j.Slf4j;
 import org.niis.xroad.cs.admin.api.domain.MemberId;
 import org.niis.xroad.cs.admin.api.facade.SignerProxyFacade;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 
 import java.util.Date;
 import java.util.List;
@@ -52,6 +59,27 @@ import java.util.List;
 @Slf4j
 @Component
 public class SignerProxyFacadeImpl implements SignerProxyFacade {
+
+    private final String signerIp;
+    private ActorSystem actorSystem;
+
+    public SignerProxyFacadeImpl(@Qualifier("signer-ip") String signerIp) {
+        this.signerIp = signerIp;
+    }
+
+    @PostConstruct
+    void init() {
+        Config config = ConfigFactory.load().getConfig("admin-service").withFallback(ConfigFactory.load());
+        actorSystem = ActorSystem.create("SignerService", config);
+        SignerClient.init(actorSystem, signerIp);
+        log.info("SignerService actorSystem initialized with admin-service config");
+    }
+
+    @PreDestroy
+    void cleanUp() {
+        actorSystem.terminate();
+    }
+
     /**
      * {@link SignerProxy#initSoftwareToken(char[])}
      */
